new: mod cli
